complain, service and user update response data

+++ src/complain/complain.service.ts
@@ -1,62 +1,55 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateComplainDto } from './dto/create-complain.dto';
-import { UpdateComplainDto } from './dto/update-complain.dto';
-import { Complain, User } from '@prisma/client';
+import { User } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { AppSuccess } from 'src/utils/AppSuccess';
 
 @Injectable()
 export class ComplainService {
+  constructor(private readonly prisma: PrismaService) {}
 
-  constructor(
-    private readonly prisma: PrismaService,
-  ) {}
-
-  async createComplain(createComplainDto: CreateComplainDto, user: User): Promise<Complain> {
-    return await this.prisma.complain.create({
+  async createComplain(
+    createComplainDto: CreateComplainDto,
+    user: User,
+  ): Promise<AppSuccess> {
+    const complain = await this.prisma.complain.create({
       data: {
         ...createComplainDto,
-        userId: user.id,
-      },
-      include: {
         client: {
-          select: {
-            user: {
-              select: {
-                id: false,
-                firstName: true,
-                lastName: true,
-                avatar: true,
-                phone: true,
-              }
-            }
-          },
-        },
-      },
-    })
-  }
-
-  async getAllComplains(): Promise<Complain[]> {
-    return await this.prisma.complain.findMany({
-      include: {
-        client: {
-          select: {
-            user: {
-              select: {
-                id: false,
-                firstName: true,
-                lastName: true,
-                avatar: true,
-                phone: true,
-              }
-            }
+          connect: {
+            id: user.id,
           },
         },
       },
     });
+
+    return new AppSuccess(complain, 'Complain created successfully');
   }
 
-  findOne(id: string): Promise<Complain> {
-    return this.prisma.complain.findUnique({
+  async getAllComplains(): Promise<AppSuccess> {
+    const complains = await this.prisma.complain.findMany({
+      include: {
+        client: {
+          select: {
+            user: {
+              select: {
+                id: false,
+                firstName: true,
+                lastName: true,
+                avatar: true,
+                phone: true,
+              },
+            },
+          },
+        },
+      },
+    });
+
+    return new AppSuccess({ complains }, 'Complains found successfully');
+  }
+
+  async findOne(id: string): Promise<AppSuccess> {
+    const complain = await this.prisma.complain.findUnique({
       where: {
         id,
       },
@@ -70,21 +63,32 @@ export class ComplainService {
                 lastName: true,
                 avatar: true,
                 phone: true,
-              }
-            }
+              },
+            },
           },
         },
       },
     });
+
+    if (!complain) throw new NotFoundException('Complain not found');
+
+    return new AppSuccess(complain, 'Complain found successfully');
   }
 
- 
-
-  remove(id: string): Promise<Complain> {
-    return this.prisma.complain.delete({
-      where: {
-        id,
-      },
+  async remove(id: string): Promise<AppSuccess> {
+    const complain = await this.prisma.complain.findUnique({
+      where: { id },
     });
+
+    if (!complain) throw new NotFoundException('Complain not found');
+
+    return new AppSuccess(
+      this.prisma.complain.delete({
+        where: {
+          id,
+        },
+      }),
+      'Complain deleted successfully',
+    );
   }
 }
